requirements.txt
aiogram>=3.4.0
SQLAlchemy>=2.0.27
aiosqlite>=0.20.0
python-dotenv>=1.0.1
aiohttp>=3.9.3
greenlet>=3.0.3